Adds dependency (Play Services) + Debug screen
Adds dependency on Google Play Services (to use GCM).
Adds classes to store and access registration ID, sender ID, etc. that are
needed for GCM to work.
Adds a GCMBroadcastReceiver that will in future listen for incoming messages.
Adds a debug screen that currently displays the registration ID for GCM.